# Conflicts:
#	src/main/java/com/ebicep/customentities/nms/CustomHorse.java
#	src/main/java/com/ebicep/warlords/Warlords.java
#	src/main/java/com/ebicep/warlords/abilties/LastStand.java
#	src/main/java/com/ebicep/warlords/database/repositories/games/pojos/DatabaseGameBase.java
#	src/main/java/com/ebicep/warlords/database/repositories/player/pojos/general/TournamentStats.java
#	src/main/java/com/ebicep/warlords/game/option/pvp/HorseOption.java
#	src/main/java/com/ebicep/warlords/player/general/SkillBoosts.java
#	src/main/java/com/ebicep/warlords/pve/DifficultyIndex.java
#	src/main/java/com/ebicep/warlords/util/bukkit/RemoveEntities.java

+++ src/main/java/com/ebicep/warlords/abilties/BloodLust.java
@@ -24,7 +24,7 @@ public class BloodLust extends AbstractAbility implements Duration {
 
     private int tickDuration = 300;
     private int damageConvertPercent = 65;
-    private int healReductionPercent = 10;
+    private float healReductionPercent = 10;
 
     public BloodLust() {
         super("Blood Lust", 0, 0, 31.32f, 20);
@@ -155,11 +155,11 @@ public class BloodLust extends AbstractAbility implements Duration {
         this.tickDuration = tickDuration;
     }
 
-    public int getHealReductionPercent() {
+    public float getHealReductionPercent() {
         return healReductionPercent;
     }
 
-    public void setHealReductionPercent(int healReductionPercent) {
+    public void setHealReductionPercent(float healReductionPercent) {
         this.healReductionPercent = healReductionPercent;
     }
 }
